properly merge things
