+++ static/app/utils/replays/createHighlightEvents.tsx
