+++ static/app/utils/replays/mergeAndSortEvents.tsx
